I checked and forwarded the code, now smooth worked

+++ Core/Inc/WS2812.h
@@ -8,6 +8,6 @@
 #ifndef INC_WS2812_H_
 #define INC_WS2812_H_
 
-void ConvertColorToTim(struct RGB_LED* LED_data, uint8_t* PwmValue);
+void ConvertColorToTim(struct RGB_LED* LED_data, uint32_t* PwmValue);
 
 #endif /* INC_WS2812_H_ */
